Fix focus signals dispatch
Problem: When splitting the current i3/Tmux layout or when gaining focus
indirectly by closing a i3/Tmux window/pane, focus events don't get
triggered. What they do is to "accumulate" and get triggered all
together when the Vim window/pane gets focus explicitly or when a key is
pressed.

Solution: When a focus signal is received, delay the 'doautocmd' command
by N milliseconds so that it can be canceled if another focus signal
comes by. This will make sure that if a batch of focus signals is
received at the same time, only the last one gets dispatched.